+++ Code/common_functions.py
@@ -1,6 +1,5 @@
 import nltk
 import re
-import utils
 
 
 class Tweet:
@@ -79,38 +78,3 @@ def parse(filename):
         if id == 5000:
             break
     return tweets, text_tweets
-
-
-def jaccard_similarity(a, b):
-    x = a.features
-    y = b.features
-    intersection_cardinality = len(set.intersection(*[set(x), set(y)]))
-    union_cardinality = len(set.union(*[set(x), set(y)]))
-    return intersection_cardinality/float(union_cardinality)
-
-
-def main():
-    tweets, text_tweets = parse('tweets_50k.txt')
-    tweets_len = len(tweets)
-    print('len(tweets) =', tweets_len)
-
-    T = utils.clustering(text_tweets, 4)
-
-    cluster_list = []
-    for i in range(len(T)):
-        cluster_list.append((i, T[i]))
-    cluster_list.sort(key=lambda x: x[1])
-
-    last_t = -1
-    for i, t in cluster_list:
-        if (t != last_t):
-            print()
-        print(t, i, tweets[i].features)
-        last_t = t
-
-    print('\nnumber of cluster', len(set(T)))
-
-
-if __name__ == "__main__":
-    # execute only if run as a script
-    main()
